fix: unit test after shifting hooks to common
