+++ src/components/common/index.ts
@@ -34,4 +34,5 @@ export * from './helpers/compareVersion'
 export * from './header/PageHeader'
 export * from './error/ErrorBar'
 export * from './DetectBottom'
-export * from './hooks'
+export * from './hooks/ClickOutside'
+export * from './hooks/FileReader'
